Broke apart classes inside files into their own files.
