+++ codetocad/adapters/blender/cad/sketch/sketch.py
@@ -3,35 +3,25 @@ from typing import List
 from uuid import uuid4
 
 from codetocad.interfaces.cad.sketch.sketch_interface import SketchInterface
-from codetocad.interfaces.cad.sketch.sketch_get import SketchGetInterface
 from codetocad.interfaces.cad.wire.wire_presets import WirePresetsInterface
-from codetocad.adapters.blender.cad.wire import Wire, WireAdd
+from codetocad.adapters.blender.cad.sketch.sketch_get import SketchGet
 from codetocad.adapters.blender.blender_actions.objects import create_object
-from codetocad.adapters.blender.blender_actions.collections import (
-    assign_object_to_collection,
-)
-
-
-class SketchGet(SketchGetInterface):
-    """Blender-specific sketch get operations."""
-
-    def __init__(self, sketch: "Sketch"):
-        self.sketch = sketch
-
-    def wire(self, i) -> Wire:
-        return self.sketch.wires[i]
 
 
 class Sketch(SketchInterface):
     """Blender implementation of SketchInterface."""
 
     def __init__(self, name: str | None = None):
+        # Import here to avoid circular imports
+        from codetocad.adapters.blender.cad.wire.wire import Wire
+        from codetocad.adapters.blender.cad.wire.wire_add import WireAdd
+
         # Blender-specific properties
         self.name = name or f"sketch_{str(uuid4())[:8]}"
         self._blender_object: bpy.types.Object | None = None
 
         # Initialize parent interface properties
-        self.wires: list[Wire] = []
+        self.wires: List[Wire] = []
         self.preset = WirePresetsInterface(Wire, self)
         self.get = SketchGet(self)
 
@@ -59,7 +49,7 @@ class Sketch(SketchInterface):
         if self._blender_object.name in bpy.context.scene.collection.objects:
             bpy.context.scene.collection.objects.unlink(self._blender_object)
 
-    def add(self, wire: Wire):
+    def add(self, wire):
         """
         Adds a Wire to the Sketch.
         """
@@ -81,6 +71,9 @@ class Sketch(SketchInterface):
     @property
     def draw(self):
         """Draw on the last wire in the sketch. If no wire exists, create a new one."""
+        from codetocad.adapters.blender.cad.wire.wire import Wire
+        from codetocad.adapters.blender.cad.wire.wire_add import WireAdd
+
         wire = self.wires[-1] if self.wires else None
 
         if not wire:
