+++ codetocad/adapters/blender/cad/vertex/vertex.py
