+++ codetocad/adapters/blender/cad/wire/wire_preset_class_property.py
@@ -1,0 +1,15 @@
+from codetocad.interfaces.cad.wire.wire_interface import (
+    _WirePresetClassPropertyInterface,
+)
+from codetocad.interfaces.cad.wire.wire_presets import WirePresetsInterface
+
+
+class _WirePresetClassProperty(_WirePresetClassPropertyInterface):
+    """Metaclass to provide a preset property for the Wire class."""
+
+    @property
+    def preset(self):
+        # Import here to avoid circular imports
+        from codetocad.adapters.blender.cad.wire.wire import Wire
+
+        return WirePresetsInterface(Wire, None)
